Add possibility to disable charger voltage check
If charger voltage is below 4300V or over 5800V, the kernel stops the
current charging process.

It was noted that sometimes, with kernel 4.14, this is not working
properly as the charging it stopped even if with 4.4 it works fine with
the same charger.

We don't have a kernel fix, but we added the possibility to disable this
check if needed: for this reason, add a flag in SettingsExtra to allow
the user to configure this behaviour and disable this check.

Note that this may be dangerous, so by default the check is enabled.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>