SettingsExtra: use and configure KeyHandler
Due to various reasons, at the moment it is easier to use the KeyHandler
in java, so switch back to java.

This is the same service used by Lineage, we keep also the configurable
preferences even if we don't have a way to change those on a UI yet. So:
- change some default values, for example we mute media by default on
  silent mode;
- vibration is triggered when a switch to vibration or silent mode is
  detected (Lineage vibrates when the normal mode is selected).

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>