config: set ro.hardware.egl
Android now (14 QPR2) loads libEGL_<ro.hardware.egl>.so.

If the property is not found, Android currently searches for libEGL_msm8998.so
that does not exist: we can set the property to fix the (boot) issue and
load libEGL_adreno.so which exists.

This was not an issue up to Android 14 QPR2.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>