build: task: ota: use a custom function to build the OTA package
This fixes an issue that was causing the previous OTA to still contains the
'test-keys' string in a lot of properties, probably because the AOSP
function to build the OTA is using a different target files path, and not
the one containing the files we just modified.

With the custom function, we can now pass the target files path.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>