Add custom OTA building script
Google is going to deprecate the single-copy partitions soon, or at least
they are going to remove support for building single-copy OTAs.

In order to avoid issues with this, create a bash script that can create
an OTA package starting from a sample package: the sample package contains
everything except the actual boot, vendor and system images file.

After configuring the environment, use the command
$ source vendor/rs/ota_build/ota_build.sh

to create an OTA that will be available in:
$ out/target/product/<product>/<product>_ota.zip

This OTA should be equivalent to the OTA built with the otapackage command,
but it requires much, much less time to be built and it also allows for
more customizations.

In fact, we can now easily customize the updater-script, for example. Right now,
we only substitute 'test-keys' with 'release-keys', which was previously
performed by the sign_target_files_apks python script (that was taking much
more time anyway, now we don't even need the target files).

Note that to simply things, the OTA does not include brotli compressed files,
instead it includes the raw vendor and system images: after compressing
the final zip file, this is ~10% bigger compared to the zip with the brotli compressed
images, but it much less complex to create with bash, so I will accept this
compromise.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>