ota_build: initialize metadata if needed
We now support mounting metadata, and we are using the logdump partition
for it.

logdump is a not-used 64M partition so we are free to use it as metadata
partition, however we cannot be sure of the initial status of the partition.

Android should be smart enough to format it if needed, but I found cases
where the partition was random data and it was not formatted automatically.
For this reason, check if the partition is already formatted in ext4 during
the OTA update, and if not, write zeros on it so that Android will format
it at first boot.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>