Add a migration.sh script to change keys without formatting /data
The initial script is provided by Lineage.

In case the OS keys change (platform, media, etc...), we need to change the
public key and certificate in the /data/system/packages.xml file.

This can be done using a script provided by Lineage, which however requires
the user to manually run it.

With this commit we make this automatic: at boot, a service is started from
init and if all conditions are met the packages.xml file is updated
accordingly.

The conditions:
- running an OS with the new keys
- being the first time running the script.

Note that, to make this works, we need to use the toolbox SELinux
context: this is a bit of an hack, so we plan to remove the migration script
in some months, after most users already updated their phone.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>