ota_build: improve migration keys file management
Instead of hard-coding the new keys and certs in the migration script,
we can update the ota_build script to automatically do this for us.

So:
- add placeholders instead of the new keys/certs in the migration script
- update the ota_build script to automatically parse the new keys
(in vendor/rs/config/security) and replace the placeholders in the migration
script with the new keys/certs
- update the migration script so that it applies the new keys only if they
are not already applied
- in case the ota_build script is not used, make sure that the migration
script does nothing (as it does not contain valid keys/certs).

The AOSP keys are still hard-coded in the migration script as 'old/previous
keys'.

With this, if the ota_build.sh script is used to create OTAs, the only step
needed to migrate keys is to replace the keys in vendor/rs/config/security.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>