config: sign the bluetooth packages with the bluetooth key
By default, Bluetooth apks are signed with a key stored in:
packages/modules/Bluetooth/android/app/certs

This is the same bluetooth key as the one stored in:
build/target/product/security

Whenever the build uses different keys, however, the Bluetooth apks continue
using the key in the packages/modules directory, while other components
are signed with the new keys.

This breaks the bluetooth functionality on Android.

To fix this, override the keys and make sure to build the bluetooth apk
with the bluetooth key.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>