config: prevent Telecomm overlays to be created
Without this, the Telecomm overlays apk is created in the vendor partition.

When Google Apps are installed, however, this apk is not deleted so there
is a conflict between the AOSP Telecomm overlay and the Google Apps overlay.

To prevent this issue to appear, just don't use overlays for the Telecomm
service.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>